feat(common)!: array type should be simple (#40)
This removes the propagating `where for<'a> &'a Self: IntoIterator<&'a T>` bound, `Index` and `IndexMut` bounds.

These have been totally replaced by `Deref<Target = [T]>` and `DerefMut` bounds, meaning that an "`Array`" must be coercible to a slice. This is a little more restrictive but **drastically** reduces code footprint and type complexity.